loading retirados, devido a conversa que tivemos e Intent init fixed, e try catch removido por correção usando o isfinishing
