+++ app/src/androidTest/java/br/com/concrete/tentacle/features/register/RegisterFragmentTest.kt
@@ -1,17 +1,11 @@
 package br.com.concrete.tentacle.features.register
 
-import android.app.Activity
-import android.app.Instrumentation
-import androidx.navigation.NavHost
 import androidx.test.espresso.Espresso.onData
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.intent.Intents.intending
-import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
-import androidx.test.espresso.matcher.ViewMatchers
+import androidx.test.espresso.matcher.ViewMatchers.withClassName
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
@@ -215,13 +209,6 @@ class RegisterFragmentTest : BaseFragmentNoActionBarNoBottomBarTest() {
             .check(matches(allOf(withText("ERROR MESSAGE."), isDisplayed())))
     }
 
-    private fun checkIfGoesToHome() {
-        Intents.init()
-        val matcher = allOf(hasComponent(NavHost::class.java.name))
-        val result = Instrumentation.ActivityResult(Activity.RESULT_OK, null)
-        intending(matcher).respondWith(result)
-    }
-
     private fun selectState() {
         onView(
             Matchers.allOf(
@@ -236,7 +223,7 @@ class RegisterFragmentTest : BaseFragmentNoActionBarNoBottomBarTest() {
             .inAdapterView(
                 Matchers.allOf(
                     withId(R.id.list),
-                    ViewMatchers.withClassName(
+                    withClassName(
                         Matchers.`is`("android.widget.LinearLayout"))
                             .childAtPosition(3)
                 )
@@ -257,7 +244,7 @@ class RegisterFragmentTest : BaseFragmentNoActionBarNoBottomBarTest() {
             .inAdapterView(
                 Matchers.allOf(
                     withId(R.id.list),
-                    ViewMatchers.withClassName(
+                    withClassName(
                         Matchers.`is`("android.widget.LinearLayout"))
                             .childAtPosition(3)
                 )
